Extract TransactionLogPartitionsSnapshot
