+++ java/statestore/src/main/java/sleeper/statestore/transactionlog/TransactionLogSnapshot.java
@@ -20,13 +20,11 @@ import org.apache.hadoop.fs.Path;
 import org.apache.parquet.hadoop.ParquetReader;
 import org.apache.parquet.hadoop.ParquetWriter;
 
-import sleeper.core.partition.Partition;
 import sleeper.core.range.RegionSerDe;
 import sleeper.core.record.Record;
 import sleeper.core.schema.Field;
 import sleeper.core.schema.Schema;
 import sleeper.core.schema.type.IntType;
-import sleeper.core.schema.type.ListType;
 import sleeper.core.schema.type.LongType;
 import sleeper.core.schema.type.StringType;
 import sleeper.core.statestore.AllReferencesToAFile;
@@ -46,7 +44,7 @@ import java.time.Instant;
 import java.util.List;
 
 public class TransactionLogSnapshot {
-    private static final Schema PARTITION_SCHEMA = initialisePartitionSchema();
+    private final TransactionLogPartitionsSnapshot partitionsSnapshot;
     private static final Schema FILE_SCHEMA = initialiseFilesSchema();
     private final RegionSerDe regionSerDe;
     private final TransactionLogStateStore store;
@@ -61,94 +59,19 @@ public class TransactionLogSnapshot {
         this.store = store;
         this.regionSerDe = new RegionSerDe(schema);
         this.configuration = configuration;
+        this.partitionsSnapshot = new TransactionLogPartitionsSnapshot(schema, store, configuration);
     }
 
-    public void save(java.nio.file.Path tempDir) throws StateStoreException {
-        saveFiles(tempDir);
-        savePartitions(tempDir);
+    void savePartitionsWithTransactionNumber(java.nio.file.Path tempDir, long lastTransactionNumber) throws StateStoreException {
+        partitionsSnapshot.save(tempDir, lastTransactionNumber);
     }
 
-    private void savePartitions(java.nio.file.Path tempDir) throws StateStoreException {
-        String path = createPartitionsPath();
-        try (ParquetWriter<Record> recordWriter = ParquetRecordWriterFactory.createParquetRecordWriter(
-                new Path(tempDir.resolve(path).toString()), PARTITION_SCHEMA, new Configuration())) {
-            for (Partition partition : store.getAllPartitions()) {
-                recordWriter.write(getRecordFromPartition(partition));
-            }
-        } catch (IOException e) {
-            throw new UncheckedIOException("Failed writing partitions", e);
-        }
+    StateStorePartitions loadPartitionsFromTransactionNumber(java.nio.file.Path tempDir, long lastTransactionNumber) throws StateStoreException {
+        return partitionsSnapshot.load(tempDir, lastTransactionNumber);
     }
 
-    public StateStorePartitions loadPartitionsFromTransactionNumber(java.nio.file.Path tempDir, long lastTransactionNumber) throws StateStoreException {
-        StateStorePartitions partitions = new StateStorePartitions();
-        try (ParquetReader<Record> reader = new ParquetRecordReader.Builder(
-                new Path(tempDir.resolve(createPartitionsPath(lastTransactionNumber)).toString()), PARTITION_SCHEMA)
-                .withConf(configuration)
-                .build();
-                ParquetReaderIterator recordReader = new ParquetReaderIterator(reader)) {
-            while (recordReader.hasNext()) {
-                partitions.put(getPartitionFromRecord(recordReader.next()));
-            }
-        } catch (IOException e) {
-            throw new StateStoreException("Failed loading partitions", e);
-        }
-        return partitions;
-    }
-
-    private Record getRecordFromPartition(Partition partition) {
-        Record record = new Record();
-        record.put("partitionId", partition.getId());
-        record.put("leafPartition", "" + partition.isLeafPartition()); // TODO Change to boolean once boolean is a supported type
-        String parentPartitionId;
-        if (null == partition.getParentPartitionId()) {
-            parentPartitionId = "null";
-        } else {
-            parentPartitionId = partition.getParentPartitionId();
-        }
-        record.put("parentPartitionId", parentPartitionId);
-        record.put("childPartitionIds", partition.getChildPartitionIds());
-        record.put("region", regionSerDe.toJson(partition.getRegion()));
-        record.put("dimension", partition.getDimension());
-        return record;
-    }
-
-    private Partition getPartitionFromRecord(Record record) {
-        Partition.Builder partitionBuilder = Partition.builder()
-                .id((String) record.get("partitionId"))
-                .leafPartition(record.get("leafPartition").equals("true"))
-                .childPartitionIds((List<String>) record.get("childPartitionIds"))
-                .region(regionSerDe.fromJson((String) record.get("region")))
-                .dimension((int) record.get("dimension"));
-        String parentPartitionId = (String) record.get("parentPartitionId");
-        if (!"null".equals(parentPartitionId)) {
-            partitionBuilder.parentPartitionId(parentPartitionId);
-        }
-        return partitionBuilder.build();
-    }
-
-    private String createPartitionsPath() throws StateStoreException {
-        return createPartitionsPath(store.getLastPartitionsTransactionNumber());
-    }
-
-    private String createPartitionsPath(long lastTransactionNumber) throws StateStoreException {
-        return "snapshots/" + lastTransactionNumber + "-partitions.parquet";
-    }
-
-    private static Schema initialisePartitionSchema() {
-        return Schema.builder()
-                .rowKeyFields(new Field("partitionId", new StringType()))
-                .valueFields(
-                        new Field("leafPartition", new StringType()),
-                        new Field("parentPartitionId", new StringType()),
-                        new Field("childPartitionIds", new ListType(new StringType())),
-                        new Field("region", new StringType()),
-                        new Field("dimension", new IntType()))
-                .build();
-    }
-
-    private void saveFiles(java.nio.file.Path tempDir) throws StateStoreException {
-        String path = createFilesPath();
+    void saveFilesWithTransactionNumber(java.nio.file.Path tempDir, long lastTransactionNumber) throws StateStoreException {
+        String path = createFilesPath(lastTransactionNumber);
         try (ParquetWriter<Record> recordWriter = ParquetRecordWriterFactory.createParquetRecordWriter(
                 new Path(tempDir.resolve(path).toString()), FILE_SCHEMA, new Configuration())) {
             for (AllReferencesToAFile file : store.getAllFilesWithMaxUnreferenced(100).getFiles()) {
@@ -192,10 +115,6 @@ public class TransactionLogSnapshot {
                 .totalReferenceCount((int) record.get("externalReferences") + internalReferences.size())
                 .lastStateStoreUpdateTime(Instant.ofEpochMilli((long) record.get("lastStateStoreUpdateTime")))
                 .build();
-    }
-
-    private String createFilesPath() throws StateStoreException {
-        return createFilesPath(store.getLastFilesTransactionNumber());
     }
 
     private String createFilesPath(long lastTransactionNumber) throws StateStoreException {
